Make one-hot vectors encoded from negative indices contain only zeros
keras.utils.to_categorical treats negative indices as indices from the end of the one-hot vector, in keeping with Python indexing semantics. However this is undesirable as negative indexes are used as a mask value and should produce
an invalid vector. This adds code to set one-hot vectors produces from negative indices to all-zeros.